+++ tests/controls/textctrltest.cpp
@@ -32,6 +32,8 @@
     #include "wx/stopwatch.h"
 #endif
 
+#include "wx/private/localeset.h"
+
 #include "textentrytest.h"
 #include "testableframe.h"
 #include "asserthelper.h"
@@ -288,7 +290,7 @@ void TextCtrlTestCase::StreamInput()
 #ifndef __WXOSX__
     {
         // Ensure we use decimal point and not a comma.
-        LocaleSetter setCLocale("C");
+        wxCLocaleSetter setCLocale;
 
         *m_text << "stringinput"
                 << 10
